cambios

+++ src/main/java/com/gymmanager/gym_manager/repository/ActividadRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.gymmanager.gym_manager.entity.Actividad;
 
-public interface ActividadRepository extends JpaRepository<Actividad, Long> {
+public interface ActividadRepository extends JpaRepository<Actividad, Integer> {
 
 }
