+++ src/main/java/com/gymmanager/gym_manager/repository/ClienteActividadRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.gymmanager.gym_manager.entity.ActividadCliente;
 
-public interface ClienteActividadRepository extends JpaRepository<ActividadCliente, Long> {
+public interface ClienteActividadRepository extends JpaRepository<ActividadCliente, Integer> {
 
 }
